hacks/mlops-on-gcp/artifacts/python/requirements.txt
google-cloud-aiplatform
google-cloud-datastore
google-cloud-pubsub
